ACCOUNT: Add Account Page and modify setup
This commit adds the basic accout page and modifies the Account Setup to
Share a styles sheet.

The commit also includes buttons that will direct to the "Sign In" and
"Create Account" pages for Debugging. This buttons will be replaced by
an automated script that will direct the user to sign in if they are not
signed in yet. The "Create Account" page will also only be accessible
from the "Sign In" page when the user is signed out.